Update create_function_qdgc_fillqdgc.sql
Added parameter for purging the whole table. If not data will be added.

+++ scripts/create_function_qdgc_fillqdgc.sql
@@ -4,7 +4,8 @@
 
 CREATE OR REPLACE FUNCTION public.qqdgc_fillqdgc(
 	area_input text,
-	qdgc_level int)
+	qdgc_level int,
+	purge int)
     RETURNS SETOF text 
     LANGUAGE 'plpgsql'
     COST 100
@@ -18,9 +19,11 @@ declare
 	lonlat text;
 	qdgc text;
 begin
-
-	drop table if exists tbl_qdgc;
-	create table  tbl_qdgc (qdgc varchar(20), area_reference varchar(50),level_qdgc int, cellsize_degrees decimal, lon_center decimal, lat_center decimal, area_km2 decimal, geom geometry);
+	
+	if purge=1 then
+		drop table if exists tbl_qdgc;
+		create table  tbl_qdgc (qdgc varchar(20), area_reference varchar(50),level_qdgc int, cellsize_degrees decimal, lon_center decimal, lat_center decimal, area_km2 decimal, geom geometry);
+	end if;
 
    	/* Establishing counter which defindes the depth of the grid cell creation. Here from 1 (1/2 degree) to 5 (1/32 degree) */
    	for counter in 1..qdgc_level loop
@@ -39,5 +42,3 @@ begin
 	
 end
 $BODY$;
-
-select qqdgc_fillqdgc('Tanzania',2)
